LoginHelper session variable "loggedIn".
Views are not supposed to call libraries, which makes it difficult for them to interact with the user's login state.
LoginHelper now creates a session variable called "loggedIn" which is true when the user is logged in.
Views can use the session data to determine whether the user is logged in, without calling the library.