Added resources view.
common/resources view will load common resources such as jquery, bootstrap, lightbox, etc.
This is due to issues where these frameworks were not loaded early enough.

The resources view will allow all the frameworks to be loaded using a single line of code and is expected to be called immediately after the SFSU Student Project banner.
NOTE: The resources view does not contain the actual code of the frameworks. It is still being called from the original views.
This allows for easy editing, organization, and allows only specific frameworks to be loaded when required.